New folder/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
Pillow==11.3.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
requests==2.32.3
python-dotenv==1.1.1
